split ParameterPool into small classes
